WIP progress on getting everything working
