+++ .gitignore
@@ -17,3 +17,4 @@ docs/crowdin.py
 *.mo
 .pypirc
 *.gz
+**/__pycache__
