+++ discord/ext/commands/view.py
@@ -4,6 +4,7 @@ The MIT License (MIT)
 Copyright (c) 2015-present Rapptz
 
 Permission is hereby granted, free of charge, to any person obtaining a
+copy of this software and associated documentationany person obtaining a
 copy of this software and associated documentation files (the "Software"),
 to deal in the Software without restriction, including without limitation
 the rights to use, copy, modify, merge, publish, distribute, sublicense,
